Notes fixed final
