+++ src/main/java/com/example/Cat.java
@@ -1,5 +1,4 @@
 package com.example;
-
 import java.util.List;
 
 public class Cat {
