+++ src/main/java/com/example/Predator.java
@@ -1,5 +1,4 @@
 package com.example;
-
 import java.util.List;
 
 public interface Predator {//Хищник
