+++ src/test/java/FelineTest.java
@@ -1,7 +1,6 @@
 import com.example.Feline;
 import org.junit.Assert;
 import org.junit.Test;
-
 import java.util.List;
 
 public class FelineTest {
